Sort risk inputs by their category-weighted scores
This breaks most of the integration tests as the scores have all
changed. I wasn't expecting this, but I'm currently trying to determine
whether that makes sense.

I would expect the values to change in the cases where category weights
are applied, but in the simpler case of defaults (i.e. a weight of 1)
that's still not quite making sense to me.

Committing for now as a checkpoint either way.